[MANTRUN-181] AttachArtifact task does not work in external Ant build file
When the run goal executes an Ant task that calls an external Ant build file, Ant creates a sub-project of the main Ant project that was created for the (parent) Ant task. However, the project's references are not passed to the sub-project, meaning that whatever Maven added as a reference to the main project is lost when calling the external build file. This notably includes the "maven.project" and "maven.project.helper" references that are needed for the 'attachartifact' built-in task to work.

This is resolved by overriding the way of creating sub-projects, so that references added to an Ant project by Maven are passed to any sub-projects.

git-svn-id: https://svn.apache.org/repos/asf/maven/plugins/trunk@1796671 13f79535-47bb-0310-9956-ffa450edef68