[MANTRUN-179] Seems impossible to use combine.* attributes with maven-antrun-plugin configuration
During creation of the temporary Ant file, the Maven combine.children and combine.self attributes of the XML plugin configuration elements should be removed. Also refactoring the code generating the Ant build file, which is only the responsibility of the writer and not the Mojo, to use proper XML serialization instead of string manipulations.

git-svn-id: https://svn.apache.org/repos/asf/maven/plugins/trunk@1799014 13f79535-47bb-0310-9956-ffa450edef68