Issue348 Log message only if trying to change maven property from ant task
It is expected that some values may be different, specifically the
special "ant.file" property can change between invocations of the antrun
plugin (e.g., changing name from pom.xml to .flattened.pom.xml in some
projects), and we do not want to issue warnings for such expected
situations. Instead, we log an INFO message when we are skipping a
value, and we log nothing for property values that are the same in Ant
and Maven.

The old behavior was to log a debug message, and new behavior is logging
an info message only for rare things that may be surprising, but are
generally innocuous and expected.